requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchmetrics>=1.3.0
pandas>=2.2.0
numpy>=1.26.0
opencv-python>=4.9.0.80
Pillow>=10.2.0
scikit-learn>=1.4.0
tqdm>=4.66.0
PyYAML>=6.0.1
tyro>=0.8.5
gradio>=4.32.0
fastapi>=0.115.0
uvicorn>=0.30.0
matplotlib>=3.8.0
onnxruntime>=1.18.0
ultralytics>=8.2.0
